Added xinitrc and awesome config to repo.
Install.sh changes:

Script now copies over/overwrites the rofi and awesome config, as well
as the xinitrc. Check line 89 for the xinitrc copy/overwrite.

The git-cloning steps for the theme and the plugin is now easier to
read. The path those would write to were given their own line.

Added 'man-pages' for installation, so even kernel programs will have
proper documentation available on the system.